[Pal/Linux-SGX] Sanitize return values of ocalls
When doing an ocall, untrusted part could return an invalid value for
example `write` could return size bigger than passed to it. This commit
adds checks for such invalid values.

+++ Pal/src/host/Linux-SGX/enclave_ocalls.h
@@ -86,7 +86,7 @@ int ocall_sleep (unsigned long * microsec);
 
 int ocall_socketpair (int domain, int type, int protocol, int sockfds[2]);
 
-int ocall_poll(struct pollfd* fds, int nfds, int64_t timeout_us);
+int ocall_poll(struct pollfd* fds, size_t nfds, int64_t timeout_us);
 
 int ocall_rename (const char * oldpath, const char * newpath);
